pyproject.toml
[project]
name = "agile-backlog"
dynamic = ["version"]
description = "Lightweight Kanban board tool for agentic development"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "agile-backlog contributors" },
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dependencies = [
    "pyyaml>=6.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
ui = [
    "nicegui>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "nicegui>=2.0.0",
]

[project.scripts]
agile-backlog = "agile_backlog.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agile_backlog = ["schema.yaml", "py.typed", "bundled_skills/**/*"]

[tool.setuptools.dynamic]
version = {attr = "agile_backlog.__version__"}

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
